# utils/swagger.yaml
openapi :  '3.0.0'
info : 
  title :  Jungle Meet Apis Documentation
  version :  '1.0.0'
  description :  This is the API documentation for JungleMeet Queuing
servers :
  - url :  http://localhost:3000/v1  
    description :  Local Server
components:
  schemas:
    User:
      type: object
      required:
        - name
        - password
        - email
      properties:
        _id:
          type: string
          description: The Auto-generated id of a post
          example: 6332d81195b55eda2a612059
        name:
          type: string
          description: name of the user
          example: Rachel
        password:
          type: string
          description: password of the user
          example: abcd1234
        email:
          type: string
          description: email of the user
          example: example@example.com
        avatar:
          type: string
          description: avatar of the user
          example: xxxxx
        createdTime:
          type: string
          description: the time of creating the user
          example: 2022-09-27T11:01:37.487Z
        bgImg:
          type: string
          description: background image of the user profile
          example: xxxxx
        role:
          type: string
          description: role of the user
          example: admin
        follower:
          type: array
          items:
            type: string
          description: follower of the user
          example: [6332d81195b55eda2a612058, 6332d81195b55eda2a612057]
        following:
          type: array
          items:
            type: string
          description: the following user of the user
          example: [6332d81195b55eda2a612053, 6332d81195b55eda2a612052]
        followingPost:
          type: array
          items:
            type: string
          description: followingPost of the user
          example: [6332d81195b55eda2a612153, 6332d81195b55eda2a612058, 6332d81195b55eda2a602057]
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
paths :
    /users/{id}:
      get:
        tags:
          - user
        summary: Find user by id
        description: Return a single user
        operationId: getUserById
        parameters:
          - name: id
            required: true
            in: path
            description: ID of user to return
            schema:
              type: string
        responses:
          '200':
            description: successful operation
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/User'
          '404':
            description: User not found
      patch:
        tags:
          - user
        summary: Update user by ID
        description: ID of user to return
        operationId: updateUser
        security:
          - cookieAuth: []
        requestBody:
          description:
            Add a new user
          content:
            application/json:
              schema:
                type: object
                required:
                  - name
                  - email
                  - avatar
                  - bgImg
                properties:
                  name:
                    type: string
                    description: name of the user
                    example: Rachel
                  email:
                    type: string
                    description: email of the user
                    example: example@example.com
                  avatar:
                    type: string
                    description: avatar of the user
                    example: xxxxx
                  bgImg:
                    type: string
                    description: background image of the user profile
                    example: xxxxx
        responses:
          '200':
            description: successful operation
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/User'
          '404':
            description: Not Found
    /users:
      post:
        tags:
          - user
        summary: Add a new user
        description: Return created user
        operationId: createUser
        requestBody:
          description:
            Add a new user
          content:
            application/json:
              schema:
                type: object
                required:
                  - name
                  - password
                  - email
                properties:
                  name:
                    type: string
                    description: name of the user
                    example: Rachel
                  password:
                    type: string
                    description: password of the user
                    example: abcd1234
                  email:
                    type: string
                    description: email of the user
                    example: example@example.com
                  avatar:
                    type: string
                    description: avatar of the user
                    example: xxxxx
                  bgImg:
                    type: string
                    description: background image of the user profile
                    example: xxxxx
        responses:
          '200':
            description: successful operation
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/User'
          '400':
            description: Bad request
      get:
        tags:
          - user
        summary: Get all users
        description: Return all users
        operationId: getAllUsers
        responses:
          '200':
            description: successful operation
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/User'
          '404':
            description: Not Found
      put:
        tags:
          - user
        summary: Update user by ID
        description: ID of user to return
        operationId: updateUser
        security:
          - cookieAuth: []
        requestBody:
          description:
            Add a new user
          content:
            application/json:
              schema:
                type: object
                required:
                  - name
                  - email
                  - avatar
                  - bgImg
                properties:
                  name:
                    type: string
                    description: name of the user
                    example: Rachel
                  email:
                    type: string
                    description: email of the user
                    example: example@example.com
                  avatar:
                    type: string
                    description: avatar of the user
                    example: xxxxx
                  bgImg:
                    type: string
                    description: background image of the user profile
                    example: xxxxx
        responses:
          '200':
            description: successful operation
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/User'
          '404':
            description: Not Found
    /users/login:
        post:
          tags:
            - user
          summary: User Log in and set cookie
          operationId: userLogIn
          requestBody:
            description:
              User log in
            content:
              application/json:
                schema:
                  type: object
                  required:
                    - email
                    - password
                  properties:
                    password:
                      type: string
                      description: password of the user
                      example: abcd1234
                    email:
                      type: string
                      description: email of the user
                      example: example@example.com
          responses:
            '200':
              description: successful operation
              content:
                application/json:
                  schema:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 'Successfully logged in'
            '401':
              description: login failed
              content:
                application/json:
                  schema:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 'Wrong password, try again'

            '404':
              description: Bad request
    /users/following:
      put:
        tags:
          - user
        summary: Follow or unfollow a user
        description: Follow or unfollow a user
        operationId: toggleFollowing
        security:
          - cookieAuth: []
        requestBody:
          description:
            Follow or unfollow a user
          content:
            application/json:
              schema:
                type: object
                required:
                  - following
                properties:
                  following:
                    type: string
                    description: id of the following/unfollowing user
                    example: 6332dxxxxxb55eda2a612057
        responses:
          '200':
            description: successful operation
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/User'
          '400':
            description: Bad request
    /users/reset:
      post:
        tags:
          - user
        summary: Reset password
        description: Reset password
        operationId: resetPassword
        security:
          - cookieAuth: []
        requestBody:
          description:
            Follow or unfollow a user
          content:
            application/json:
              schema:
                type: object
                required:
                    - newPassword
                    - oldPassword
                properties:
                    newPassword:
                      type: string
                      description:  new password of the user
                      example: abcd5678
                    oldPassword:
                      type: string
                      description: old password of the user
                      example: abcd1234
        responses:
          '200':
            description: successful operation
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/User'
          '400':
            description: Bad request
...